pyproject.toml
[project]
name = "sa-pharma-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastmcp>=2.14.2",
    "httpx[http2]>=0.28.1",
    "pandas>=2.3.3",
]
